refactor(ui-common): remove DelayedCircularProgressIndicator and hide modifier to onboarding permissions module
